Create separate profile for JAR signin
Taken from the official Sonatype blog (https://blog.sonatype.com/2010/01/how-to-generate-pgp-signatures-with-maven/):

Usually when you build SNAPSHOT version, you don't want to sign the artifacts, because it's unnecessary and time-consuming. Artifact signing is only necessary when it's time to release project. So, we can wrap the plugin configuration with a profile, which will only be activated on project releasing.