Fix-up some changes made during rebasing and extend tests
This is all in one commit because the change and the tests are related.
I originally made some minor changes during the rebasing to some commits
in order to support changes made to the data structures in the
Javascript.

I extended the tests to make sure that hidden elements are handled
correctly.